Started breaking out controller into separate objects
